+++ models/area.go
@@ -8,10 +8,9 @@ type Area struct {
 	Name           string      `gorm:"not null" json:"name"`
 	Province       string      `gorm:"not null" json:"province"`
 	Signature      string      `json:"signature"`
-	CodeEntreprise uint        `json:"code_entreprise"`
+	CodeEntreprise uint64      `json:"code_entreprise"`
 	Livraisons     []Livraison `gorm:"foreignKey:AreaID"`
 }
-
 
 type AreaCount struct {
 	AreaName string `json:"area_name"`
